+++ Source/GPU/Add On - DirectX/CDirectXShaderBuiltIns.cpp
@@ -71,7 +71,8 @@ class CDirectXVertexShaderBasic : public CDirectXVertexShader {
 															};
 						HRESULT	result =
 										d3dDevice.CreateInputLayout(vertexDesc, ARRAYSIZE(vertexDesc),
-												shaderData.getBytePtr(), shaderData.getByteCount(), &mInputLayout);
+												shaderData.getBytePtr(), (SIZE_T) shaderData.getByteCount(),
+												&mInputLayout);
 						AssertFailIf(FAILED(result));
 
 						// Create constant buffer
@@ -143,7 +144,8 @@ class CDirectXVertexShaderClip : public CDirectXVertexShader {
 															};
 						HRESULT	result =
 										d3dDevice.CreateInputLayout(vertexDesc, ARRAYSIZE(vertexDesc),
-												shaderData.getBytePtr(), shaderData.getByteCount(), &mInputLayout);
+												shaderData.getBytePtr(), (SIZE_T) shaderData.getByteCount(),
+												&mInputLayout);
 						AssertFailIf(FAILED(result));
 
 						// Create constant buffer
